+++ Task.h
@@ -7,13 +7,13 @@ private:
 	unsigned id;
 	Product	requiredProduct;
 	unsigned requiredQuantity;
-	int rewardBalance;
+	double rewardBalance;
 	int	rewardScore;
 public:
 	Task(const Product& requiredProduct, unsigned requiredQuantity, int rewardBalance, int rewardScore);
 	const unsigned& getId() const;
 	const Product& getRequiredProduct() const;
 	const unsigned& getRequiredQuantity() const;
-	const int& getRewardBalance() const;
+	const double& getRewardBalance() const;
 	const int& getRewardScore() const;
 };
